ci(test): add ci/cd

+++ providers/s3_storage_provider.py
@@ -13,15 +13,11 @@ class AwsS3Provider(StorageInterface):
         'aws_secret_access_key': os.getenv('aws_secret_access_key'),
         }
         self.bucket_name = 'interfaces-prac' 
-        try:       
-            self.s3_resource = boto3.resource('s3',**credentials)
-            self.s3_client = boto3.client('s3', **credentials)   
-            self.bucket = self.s3_resource.Bucket(name= self.bucket_name)
-            return (True, "Connected Successfully")
-
-        except(Exception) as err:
-            print(err)
-            return (False, "Error")
+   
+        self.s3_resource = boto3.resource('s3',**credentials)
+        self.s3_client = boto3.client('s3', **credentials)   
+        self.bucket = self.s3_resource.Bucket(name= self.bucket_name)
+        return (True, "Connected Successfully")
 
     def disconnect(self):
         return (True, "Disconnected Successfully")
@@ -29,9 +25,9 @@ class AwsS3Provider(StorageInterface):
     def upload_file(self, source_uri: str, dest_url: str):
         try:
             response = self.s3_client.upload_file(source_uri, self.bucket_name, dest_url)
-            return (True,"Uploaded Successfully")
+            return (True, 'File transferred')
 
-        except(ClientError) as e:
+        except(Exception, ClientError) as e:
             print(e)
             return (False, "Error")
 
@@ -39,14 +35,16 @@ class AwsS3Provider(StorageInterface):
         dl_dest = dest_url
         try:
             response = self.s3_client.download_file(self.bucket_name, source_uri, dl_dest)
-            return (True, "Downloaded Successfully")
+            return (True, 'File transferred')
 
-        except(ClientError) as e:
+        except(Exception, ClientError) as e:
             print(e)
             return (False,"Error")
         
     def get_secure_file_url(self, path: str):
         try:
+            if path is None:
+                raise Exception()
             response = self.s3_client.generate_presigned_url('get_object',
                                                     Params={'Bucket': self.bucket_name,
                                                             'Key': path},
@@ -54,15 +52,17 @@ class AwsS3Provider(StorageInterface):
             # The response contains the presigned URL
             return (True, response)
 
-        except ClientError as e:
+        except (Exception, ClientError) as e:
             print(e)
             return (False, "Error")
 
     def delete_file(self, path: str):
         try:
+            if path is None:
+                raise Exception()
             self.s3_client.delete_object(Bucket=self.bucket_name, Key=path)
             print(f'File at {path} deleted succesfully')
-            return (True, "Deleted Successfully")
+            return (True, "File deleted")
         
         except(Exception, ClientError) as err:
             print(f'Error: {err}')
@@ -85,6 +85,8 @@ class AwsS3Provider(StorageInterface):
     def list_files_in_directory(self, dir: str):
         files = []
         try:
+            if dir is None:
+                raise Exception()
             for obj in self.bucket.objects.all():
                 if obj.key.__contains__(dir):
                     files.append(obj.key)
@@ -103,7 +105,7 @@ class AwsS3Provider(StorageInterface):
                     return (True, 'File exists')
             
             print(f'File at {file_uri} doesnot exist')
-            return (False, 'File doesnot exists')
+            return (False, 'Error')
 
         except(Exception, ClientError) as err:
             print(f'Error: {err}')
